Added backwards compatibility to v5 on Object/Get with AccessPoint

+++ src/app/CHAOS.MCM/Extension/Mcm.cs
@@ -14,7 +14,7 @@
 
         #endregion
 
-        public Trace Index(string view)
+        public Trace Index(string view, uint? folderId, bool cleanIndex = false)
         {
             var deleteStopwatch    = new System.Diagnostics.Stopwatch();
             var objectGetStopwatch = new System.Diagnostics.Stopwatch();
@@ -22,20 +22,22 @@
             var totalCount         = 0;
 
             deleteStopwatch.Start();
-            
-            if(string.IsNullOrEmpty(view)) 
-                ViewManager.Delete();
-            else
-                ViewManager.GetView(view).Delete();
+
+            if (cleanIndex){
+                if(string.IsNullOrEmpty(view)) 
+                    ViewManager.Delete();
+                else
+                    ViewManager.GetView(view).Delete();
+            }
             
             deleteStopwatch.Stop();
 
-            const uint PageSize = 1000;
+            const uint PageSize = 500;
             
             for (uint i = 0; ; i++)
             {
                 objectGetStopwatch.Start();
-                var objects = McmRepository.ObjectGet(null, i, PageSize, true, true, true, true, true);
+                var objects = McmRepository.ObjectGet(folderId, i, PageSize, true, true, true, true, true);
                 objectGetStopwatch.Stop();
 
                 indexStopwatch.Start();
